Improve the Internal menu
