+++ Moments/Moments/Features/InternalMenu/Views/InternalMenuActionTriggerCell.swift
@@ -11,10 +11,5 @@ class InternalMenuActionTriggerCell : InternalMenuCell<InternalMenuActionTrigger
     override func update(_ item: InternalMenuActionTriggerItemViewModel) {
         accessoryType = .disclosureIndicator
         textLabel?.text = item.title
-        textLabel?.textColor = item.titleColor
-        textLabel?.textAlignment = item.textAlignment
-        detailTextLabel?.attributedText = item.detailAttributedText
-        imageView?.image = item.icon
-        imageView?.tintColor = item.iconTint
     }
 }
